requirements.txt
pymongo[srv]
discord.py
paypalrestsdk